Tue, Aug 15, 2023, 5:42 PM +09:00

+++ readability/readability.c
@@ -85,8 +85,8 @@ int count_sentences(string text)
 
 void reading_level(int c, int w, int s)
 {
-    double L = c / w * 100;
-    double S = s / w * 100;
+    double L = c / (float)w * 100;
+    double S = s / (float)w * 100;
     double index = (0.0588 * L - 0.296 * S - 15.8);
     //printf("CALCULATED INDEX: %f\n", index);
     if (index < 2)
